fix: redundant code
