+++ internal/utils/mailing/send_mail.go
@@ -1,14 +1,34 @@
 package mailing
 
 import (
-	emailconf "Go-Starter-Template/internal/config/email_config"
+	"Go-Starter-Template/internal/utils"
 	"strconv"
 
 	"gopkg.in/gomail.v2"
 )
 
+type MailConfig struct {
+	AppURL       string
+	SMTPHost     string
+	SMTPPort     string
+	SMTPSender   string
+	SMTPEmail    string
+	SMTPPassword string
+}
+
+func LoadMailConfig() MailConfig {
+	return MailConfig{
+		AppURL:       utils.GetEnv("APP_URL"),
+		SMTPHost:     utils.GetEnv("SMTP_HOST"),
+		SMTPPort:     utils.GetEnv("SMTP_PORT"),
+		SMTPSender:   utils.GetEnv("SMTP_SENDER_NAME"),
+		SMTPEmail:    utils.GetEnv("SMTP_AUTH_EMAIL"),
+		SMTPPassword: utils.GetEnv("SMTP_AUTH_PASSWORD"),
+	}
+}
+
 func SendMail(toEmail string, subject string, body string) error {
-	emailConfig := emailconf.LoadMailConfig()
+	emailConfig := LoadMailConfig()
 
 	mailer := gomail.NewMessage()
 	mailer.SetHeader("From", emailConfig.SMTPEmail)
